requirements.txt
streamlit>=1.33.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
plotly>=5.18.0
pandas>=2.0.0
scikit-learn>=1.3.0